Move the type parser to a dedicated translation unit
